api: update 209

+++ runelite-api/src/main/java/net/runelite/api/Scene.java
@@ -67,13 +67,13 @@ public interface Scene
 	 * Get the underlay ids for the scene. The value stored is id + 1, with 0 for no underlay.
 	 * @return
 	 */
-	byte[][][] getUnderlayIds();
+	short[][][] getUnderlayIds();
 
 	/**
 	 * Get the overlay ids for the scene. The value stored is id + 1, with 0 for no overlay.
 	 * @return
 	 */
-	byte[][][] getOverlayIds();
+	short[][][] getOverlayIds();
 
 	/**
 	 * Get the shapes of the tiles for the scene.
